od_ft_tt

This object contains the trips made from origin (TAZ) to destination (TAZ) by full-time workers to work destinations; it is sourced from the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area. Also contained within is the calculated car travel time for each trip assuming a 7:00 am departure from the centroid of the TAZ on Oct. 20 2021.

data(od_ft_tt)

Format

A dataframe containing 103076 rows and 4 variables; each row represents a unique trip from origin TAZ to destination TAZ with associated features.

Origin

The unique ID of the origin traffic analysis zone (TAZ).

Destination

The unique ID of the destination traffic analysis zone (TAZ).

Persons

The number of people making this trip.

travel_time

The car travel time from origin to destination assuming a 7:00 am departure from and to the TAZ centroids on Oct. 20 2021.

Source

"2016 Transportation Tomorrow Survey" from the Data Management Group accessed November 14 2021. Travel times calculated using r5r